LS/NFFT: own psi tables for the w grid, FLT-typed floor() in fast_gaussian_grid (defect 3, nfft-psi-table + nfft-floorf-double)
Root cause (a): LombScargleMemory.allocate_grids gave the w-spectrum NFFT
memory precomp_psi=False and the yw memory's q1/q2/q3 tables while sizing
the two grids differently (ng_w ~ 2 ng_yw). precompute_psi encodes
frac(ng * x) for its own grid length, so every point's Gaussian window on
the w grid was displaced by frac(ng_yw x) - frac(ng_w x) cells: the
w-spectrum was off by 0.15 (exact-DFT comparison) and every default-path
Lomb-Scargle power was biased by 3e-3..2.4e-2, identically in float32 and
float64 and independent of m. Fix: each NFFT memory allocates and
precomputes its own tables.

Root cause (b): cunfft.cu fast_gaussian_grid used floorf() on the grid
coordinate, which under DOUBLE_PRECISION rounds the double to float32
before flooring; points within a float32 ulp below an integer were
deposited one cell right of the window precompute_psi centred (exact
fraction), so use_double=True was LESS accurate than float32 on dense
grids (2.3e-3 vs 1.8e-3 vs astropy on a 1000-point, 3-yr, 109K-frequency
grid). Fix: floor() (FLT overload; identical to floorf for float32 --
float32 results are bit-identical).

Default-path results change: every NFFT-path Lomb-Scargle power moves by
up to ~1e-2 toward the exact GLS (A40: k0=1 grid, N=300, T=365 d: float32
7.7e-3 -> 1.9e-4, double 7.7e-3 -> 4.3e-8 vs astropy; long-baseline grid
double 2.3e-3 -> 2.0e-8, float32 1.8e-3 -> 1.6e-4). The floor() fix
changes only use_double=True.

Tests: test_lombscargle.py tolerances tightened 1e-2 -> 1e-4 (the old code
fails at that level); new TestLombScargleAccuracy (float32 <= 6e-4 and
double <= 1e-6 vs astropy on the k0=1 and long-baseline grids; device
w/yw spectra vs the exact float64 DFT); test_nfft.py
test_fast_grid_double_precision_floor (deterministic ng*x - m = 16 - 2^-30
case: window must start at cell 15).

+++ cuvarbase/kernels/cunfft.cu
@@ -147,8 +147,15 @@ __global__ void fast_gaussian_grid(
 		// observation
 		FLT yi = y[i];
 
-		// nearest gridpoint (rounding down)
-		int u = (int) floorf(ng * xval - m);
+		// nearest gridpoint (rounding down). Must be the FLT-typed
+		// floor(): under DOUBLE_PRECISION floorf() rounded the double
+		// coordinate to float32 first, so points within a float32 ulp
+		// below an integer were deposited one cell to the right of
+		// where precompute_psi (which uses the exact fraction) placed
+		// the window -- ~n0*ng/2^24 misplaced points, making
+		// use_double=True LESS accurate than float32 at survey scale
+		// (nfft-floorf-double, Sep 2026). For float, floor() is floorf().
+		int u = (int) floor(ng * xval - m);
 
 		// precomputed filter values
 		FLT Q  = q1[di];
